+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add R-like functions for fitting linear models in module `utils`
 - add a function to compute natural cubic spines basis in module `utils`
 - add tmixture functions from R package limma
+- add linear model fit functions from R package limma
 
 ## [0.4.1]